requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-decouple==2.1
django-filter==24.2
djangorestframework==3.15.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Markdown==3.6
pika==1.3.2
pyTelegramBotAPI==4.18.1
python-telegram-bot==21.2
requests==2.31.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
